+++ client/js/src/client.ts
@@ -150,44 +150,52 @@ export class Client {
 				await this.resolve_cookies();
 			}
 
-			await this._resolve_config().then(async ({ config }) => {
-				if (config) {
-					this.config = config;
-					if (this.config && this.config.connect_heartbeat) {
-						if (this.config.space_id && this.options.hf_token) {
-							this.jwt = await get_jwt(
-								this.config.space_id,
-								this.options.hf_token,
-								this.cookies
-							);
-						}
-					}
-				}
-
-				if (config.space_id && this.options.hf_token) {
-					this.jwt = await get_jwt(config.space_id, this.options.hf_token);
-				}
-
-				if (this.config && this.config.connect_heartbeat) {
-					// connect to the heartbeat endpoint via GET request
-					const heartbeat_url = new URL(
-						`${this.config.root}/heartbeat/${this.session_hash}`
-					);
-
-					// if the jwt is available, add it to the query params
-					if (this.jwt) {
-						heartbeat_url.searchParams.set("__sign", this.jwt);
-					}
-
-					this.heartbeat_event = await this.stream(heartbeat_url); // Just connect to the endpoint without parsing the response. Ref: https://github.com/gradio-app/gradio/pull/7974#discussion_r1557717540
-				}
-			});
+			await this._resolve_config().then(({ config }) =>
+				this._resolve_hearbeat(config)
+			);
 		} catch (e: any) {
 			throw Error(e);
 		}
 
 		this.api_info = await this.view_api();
 		this.api_map = map_names_to_ids(this.config?.dependencies || []);
+	}
+
+	async _resolve_hearbeat(_config: Config): Promise<void> {
+		if (_config) {
+			this.config = _config;
+			if (this.config && this.config.connect_heartbeat) {
+				if (this.config.space_id && this.options.hf_token) {
+					this.jwt = await get_jwt(
+						this.config.space_id,
+						this.options.hf_token,
+						this.cookies
+					);
+				}
+			}
+		}
+
+		if (_config.space_id && this.options.hf_token) {
+			this.jwt = await get_jwt(_config.space_id, this.options.hf_token);
+		}
+
+		if (this.config && this.config.connect_heartbeat) {
+			// connect to the heartbeat endpoint via GET request
+			const heartbeat_url = new URL(
+				`${this.config.root}/heartbeat/${this.session_hash}`
+			);
+
+			// if the jwt is available, add it to the query params
+			if (this.jwt) {
+				heartbeat_url.searchParams.set("__sign", this.jwt);
+			}
+
+			// Just connect to the endpoint without parsing the response. Ref: https://github.com/gradio-app/gradio/pull/7974#discussion_r1557717540
+			if (!this.heartbeat_event)
+				this.heartbeat_event = await this.stream(heartbeat_url);
+		} else {
+			this.heartbeat_event?.close();
+		}
 	}
 
 	static async connect(