+++ js/_website/src/routes/[[version]]/docs/python-client/+page.server.ts
@@ -2,7 +2,7 @@ import { redirect } from "@sveltejs/kit";
 
 export function load({ params }) {
 	if (params?.version)
-		throw redirect(302, `/${params?.version}/docs/python-client/intro`);
+		throw redirect(302, `/${params?.version}/docs/python-client/introduction`);
 
-	throw redirect(302, `/docs/python-client/intro`);
+	throw redirect(302, `/docs/python-client/introduction`);
 }