+++ js/row/Index.svelte
@@ -1,9 +1,16 @@
 <script lang="ts">
+	import { StatusTracker } from "@gradio/statustracker";
+	import type { LoadingStatus } from "@gradio/statustracker";
+	import type { Gradio } from "@gradio/utils";
+
 	export let equal_height = true;
 	export let elem_id: string;
 	export let elem_classes: string[] = [];
 	export let visible = true;
 	export let variant: "default" | "panel" | "compact" = "default";
+	export let loading_status: LoadingStatus | undefined = undefined;
+	export let gradio: Gradio | undefined = undefined;
+	export let show_progress = false;
 </script>
 
 <div
@@ -15,6 +22,18 @@
 	id={elem_id}
 	class={elem_classes.join(" ")}
 >
+	{#if loading_status && show_progress && gradio}
+		<StatusTracker
+			autoscroll={gradio.autoscroll}
+			i18n={gradio.i18n}
+			{...loading_status}
+			status={loading_status
+				? loading_status.status == "pending"
+					? "generating"
+					: loading_status.status
+				: null}
+		/>
+	{/if}
 	<slot />
 </div>
 
@@ -24,6 +43,7 @@
 		flex-wrap: wrap;
 		gap: var(--layout-gap);
 		width: var(--size-full);
+		position: relative;
 	}
 
 	.hide {