+++ js/json/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @gradio/json
 
+## 0.2.6
+
+### Dependency updates
+
+- @gradio/statustracker@0.6.0
+
 ## 0.2.5
 
 ### Dependency updates