+++ js/multimodaltextbox/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @gradio/multimodaltextbox
 
+## 0.4.7
+
+### Dependency updates
+
+- @gradio/upload@0.10.7
+- @gradio/client@0.20.1
+- @gradio/image@0.11.7
+
+## 0.4.6
+
+### Dependency updates
+
+- @gradio/client@0.20.0
+- @gradio/statustracker@0.6.0
+- @gradio/image@0.11.6
+- @gradio/upload@0.10.6
+
 ## 0.4.5
 
 ### Dependency updates