+++ js/simpleimage/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @gradio/simpleimage
 
+## 0.5.7
+
+### Dependency updates
+
+- @gradio/upload@0.10.7
+- @gradio/client@0.20.1
+
+## 0.5.6
+
+### Dependency updates
+
+- @gradio/client@0.20.0
+- @gradio/statustracker@0.6.0
+- @gradio/upload@0.10.6
+
 ## 0.5.5
 
 ### Dependency updates