+++ js/upload/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @gradio/upload
 
+## 0.10.7
+
+### Dependency updates
+
+- @gradio/upload@0.10.7
+- @gradio/client@0.20.1
+
+## 0.10.6
+
+### Dependency updates
+
+- @gradio/client@0.20.0
+- @gradio/upload@0.10.6
+
 ## 0.10.5
 
 ### Dependency updates